refactor: 기존 fitness 패키지 분리
- 기존 fitness 패키지를 한 번 더 들어가도록 수정
