+++ src/main/java/com/flab/eattofit/exercise/domain/fitness/fitness/FitnessRepository.java
@@ -1,4 +1,4 @@
-package com.flab.eattofit.exercise.domain.fitness;
+package com.flab.eattofit.exercise.domain.fitness.fitness;
 
 public interface FitnessRepository {
