+++ src/main/java/com/flab/eattofit/exercise/ui/fitness/FitnessController.java
@@ -2,7 +2,7 @@ package com.flab.eattofit.exercise.ui.fitness;
 
 import com.flab.eattofit.exercise.application.fitness.FitnessService;
 import com.flab.eattofit.exercise.application.fitness.dto.FitnessCreateRequest;
-import com.flab.eattofit.exercise.domain.fitness.Fitness;
+import com.flab.eattofit.exercise.domain.fitness.fitness.Fitness;
 import com.flab.eattofit.exercise.ui.fitness.dto.FitnessCreateResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
